move userpic component and refactor with new image component
